requirements.txt
python-dateutil==2.7.5
faker
google
flask
pytest
pylint
-e git+https://gitlab.com/pycqa/flake8#egg=flake8
gunicorn
flask_pymongo
pymongo==3.4.0
requests>=2.20.0
passlib[totp]
python-coveralls
coverage
pytest-cov
google-auth
flask-cors==3.0.6
haversine
PyJWT
